Configured models, code review etc "

+++ api/users/routes.py
@@ -4,9 +4,9 @@ from datetime import datetime, timedelta
 import jwt
 
 from api import db, bcrypt
-from api.models import User, Post
-from api.schema import user_input, user_response, PostResponse, user_login_input, user_update_input, reset_user_password
-from api.utils import generate_hash_password, token_required, get_reset_token, verify_reset_token, send_email , get_user_by_email, get_user_by_id, serialize_user, get_serialized_user_by_id
+from api.models import User
+from api.schema import user_register, user_response, user_login, user_update, reset_password
+from api.utils import generate_hash_password, token_required, admin_token_required,get_reset_token, verify_reset_token, send_reset_password_email , get_user_by_email, get_user_by_id,  get_serialized_user_by_id, get_verification_token, send_verify_email
 
 users = Blueprint("users", __name__,)
        
@@ -15,36 +15,32 @@ def register():
     data = request.get_json()
     
     try:
-        data = user_input.load(data)  
-        user = get_user_by_email(data["email"]) 
-        
-        if user :
-            return {"message": "The User with this Email already exist."}, 409
-            
-        data["password"] = generate_hash_password(data["password"])
-        user = User(**data)
-        
-        db.session.add(user)
-        db.session.commit()
-        
+        data = user_register.load(data)  
     except ValidationError as err:
         return {"message" : err.messages}, 400
     
+    user = get_user_by_email(data["email"]) 
+        
+    if user :
+        return {"message": "The User with this Email already exist."}, 409
+            
+    data["password"] = generate_hash_password(data["password"])
+    user = User(**data)
     
-    # token = jwt.encode({"id": user.id, "exp": datetime.utcnow() + timedelta(minutes= 30 )}, current_app.config["SECRET_KEY"])
-    # email_sent = send_email(user, token)
-    # if not email_sent:
-    #     return {"message" : "Due to some reason your email couldnot be valiated right now. Please Validate your email after some time."}, 401
-    # return {"message": "The validation link is sent to your email. Please go and confirm it."}, 201
+    db.session.add(user)
+    db.session.commit()
+    
+    token = get_verification_token(user.id)
+    send_verify_email(user.email, token)
+
     return {"message": "User has been Created"}, 201
         
-    
 @users.route("/login", methods = ["POST"])
 def login():
     auth = request.authorization
  
     try:
-        data = user_login_input.load(auth)
+        data = user_login.load(auth)
     except ValidationError as err:
         return {"message" : err.messages}, 400
     
@@ -60,56 +56,55 @@ def login():
     
     return {"token" : token}
 
-
 @users.route("/account/<int:user_id>", methods=["GET"])
-def user_account(user_id):
+def get_user_account(user_id):
+    
     user = User.query.get(user_id)
+    
     if not user :
-        return {"message": "User doesnot exist in the database"}, 404
+        return {"message": "User doesnot exist."}, 404
     
     user = user_response.dump(user)
-    posts = Post.query.filter_by(user_id = user["id"]).all()
-    user["posts"]= PostResponse(exclude=["user"]).dump(posts, many=True)
-   
+    
     return {"user": user}, 200
 
 @users.route("/update_user", methods=["PUT"])
 @token_required
 def update_user(user):
+    
     data = request.get_json()
+    
     try:
-        data = user_update_input.load(data)
-        user_exist = get_user_by_email(data["email"])
-        
-        if user_exist :
-            return {"message": "The Email already exist."}, 400
-        
+        data = user_update.load(data)
     except ValidationError as err:
         return {"message" : err.messages}, 400
-
-    user.email = data["email"]
-    user.username = data["username"]
-    user.password=  generate_hash_password(data["password"])
     
-    user = serialize_user(user)
+    user.username = data["username"]
+
     db.session.commit()
     
-    return {"user" : user}, 200
+    return {"user" : "User has been Updated"}, 200
 
-
-@users.route("/delete_user", methods=["DELETE"])
-@token_required
-def delete_user(user):
+@users.route("/delete_user/<int:user_id>", methods=["DELETE"])
+@admin_token_required
+def delete_user(user, user_id):
+     
+    user = User.query.get(user_id)  
+    if not user :
+        return {"message": "The User doesnot Exist."}, 404
+    
     db.session.delete(user)
     db.session.commit()
+    
     return {"message" : "The user has been deleted"}, 200
     
-
 @users.route("/generate_token")
-def generate_token():    
+def generate_reset_password_token():    
+    
     email = request.args.get("email")
+    
     if email is None :
-        return {"message": "Username field is Missing."}, 400
+        return {"message": "Email field is Missing."}, 400
     
     user = get_user_by_email(email)
     
@@ -117,28 +112,38 @@ def generate_token():
         return {"message" : "User doesnot exist."}, 400
     
     token = get_reset_token(user.id)
-    send_email(user, token)
-    return {"message" : "Token Sent"}
+    send_reset_password_email(user.email, token)
+    
+    return {"message" : "Token Sent"}, 200
     
 @users.route("/verify_token/<token>", methods=["GET"])
 def verify_token(token):
-    token_data = verify_reset_token(token)
-    if not token_data :
+    
+    user = verify_reset_token(token)
+    
+    if not user :
         return {"message" : "The Token is Invalid or Expired"}, 401
     
-    user = get_serialized_user_by_id(token_data["user_id"])
+    user = get_serialized_user_by_id(user["user_id"])
+    
     return {"user": user}, 200
     
 @users.route("/reset_password/", methods=["POST"])
 def reset_passsword():
+    
     data = request.get_json()
+    
     try:    
-        user_credentials = reset_user_password.load(data)
+        user_credentials = reset_password.load(data)
     except ValidationError as err:
         return {"message" : err.messages}, 401
     
     user = get_user_by_id(user_credentials["user_id"])
+    if not user:
+        return {"message": "The user doesnot exist"}, 404
+    
     user.password = generate_hash_password( user_credentials["password"])
+    
     db.session.commit()
     
     return {"message" : "The password has been updated."}, 200
